srv: Only build server components on Linux
It's not clear that running the server component on Mac would have
ever worked. Now it will not even compile due to the container
networking ns package.
